+++ crates/nostr-sdk/src/lib.rs
@@ -29,9 +29,9 @@ pub use nostr_lmdb::NostrLMDB;
 pub use nostr_ndb::{self as ndb, NdbDatabase};
 #[doc(hidden)]
 pub use nostr_relay_pool::{
-    self as pool, AtomicRelayServiceFlags, FilterOptions, Relay, RelayConnectionStats,
-    RelayOptions, RelayPool, RelayPoolNotification, RelayPoolOptions, RelayServiceFlags,
-    RelayStatus, SubscribeAutoCloseOptions, SubscribeOptions, SyncDirection, SyncOptions,
+    self as pool, AtomicRelayServiceFlags, Relay, RelayConnectionStats, RelayOptions, RelayPool,
+    RelayPoolNotification, RelayPoolOptions, RelayServiceFlags, RelayStatus,
+    SubscribeAutoCloseOptions, SubscribeOptions, SyncDirection, SyncOptions,
 };
 #[doc(hidden)]
 #[cfg(feature = "nip57")]